backend/requirements.txt
# Nectar & Nut Backend Dependencies
# Worker: Dev - Python package requirements

# Django Core
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0

# Authentication
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.61.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Payment Gateway
razorpay>=1.4.0

# Email Service
resend>=2.0.0

# Report Generation
pandas>=2.2.0
reportlab>=4.1.0
openpyxl>=3.1.0

# Image Processing
Pillow>=10.2.0

# Environment Variables
python-decouple>=3.8

# Development
ipython>=8.0.0

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
coverage>=7.4.0
django-filter==25.2
